server/requirements.txt
flask
flask-cors
requests
werkzeug
apscheduler
cryptography
websocket-client
